requirements.txt
psycopg2-binary
pytest
testcontainers
openpyxl>=3.1,<4.0
